AbstractStringBuffer 相关代码移动至 text 下

+++ src/native/Windows/system/SharedMemory.cpp
@@ -8,7 +8,7 @@ SharedMemory::Ptr SharedMemory::create(const char *name, size_t size) noexcept {
             nullptr,
             PAGE_READWRITE,
             0,
-            size,
+            (DWORD) size,
             name
     );
